pwm: Fix some arches by using %ju and casting to uintmax_t
Reported by:	ci.freebsd.org


git-svn-id: svn+ssh://svn.freebsd.org/base/head@342007 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.sbin/pwm/pwm.c
@@ -190,9 +190,9 @@ main(int argc, char *argv[])
 		}
 		break;
 	case PWM_SHOW_CONFIG:
-		printf("period: %lu\nduty: %lu\nenabled:%d\n",
-		    state.period,
-		    state.duty,
+		printf("period: %ju\nduty: %ju\nenabled:%d\n",
+		    (uintmax_t) state.period,
+		    (uintmax_t) state.duty,
 		    state.enable);
 		break;
 	case PWM_PERIOD:
